requirements.txt
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
psycopg2==2.9.6
requests>=2.31.0
responses==0.23.3
slack-bolt
flake8==6.0.0
black==22.10.0
aiohttp